# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AI Safety Incident API",
    "version": "1.0.0",
    "description": "API for managing AI safety incidents and reports",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "schemas": {
      "Incident": {
        "type": "object",
        "required": [
          "title",
          "description",
          "severity"
        ],
        "properties": {
          "_id": {
            "type": "string",
            "description": "Auto-generated MongoDB ID"
          },
          "title": {
            "type": "string",
            "description": "Title of the incident"
          },
          "description": {
            "type": "string",
            "description": "Detailed description of the incident"
          },
          "severity": {
            "type": "string",
            "enum": [
              "Low",
              "Medium",
              "High"
            ],
            "description": "Severity level of the incident"
          },
          "reported_at": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the incident was reported"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "error": {
            "type": "object"
          }
        }
      }
    }
  },
  "paths": {
    "/api/incidents": {
      "post": {
        "summary": "Create a new incident",
        "tags": [
          "Incidents"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "title",
                  "description",
                  "severity"
                ],
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "severity": {
                    "type": "string",
                    "enum": [
                      "Low",
                      "Medium",
                      "High"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created incident",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Incident"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/incidents/{id}": {
      "get": {
        "summary": "Get an incident by ID",
        "tags": [
          "Incidents"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Incident ID"
          }
        ],
        "responses": {
          "200": {
            "description": "The incident",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Incident"
                }
              }
            }
          },
          "404": {
            "description": "Incident not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete an incident",
        "tags": [
          "Incidents"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Incident ID"
          }
        ],
        "responses": {
          "204": {
            "description": "Incident deleted successfully"
          },
          "404": {
            "description": "Incident not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "tags": []
}